# OpenApiSpec.yml
openapi: "3.0.0"
info:
  title: "express-api"
  version: "1.2.4"
servers:
  - url: "https://localhost/api/v1"
    description: "Local development server"
  - url: "http://localhost:8888/api/v1"
    description: "Unsecured Local development server"
  - url: "https://api.example.com/api/v1"
    description: "Production server"
paths:
...
  /user:
    get:
      description: "Get the current user"
      tags:
        - "User"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Returns the current user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: "Unauthorized - Invalid or missing authentication token"
    post:
      description: "Update the current user"
      tags:
        - "User"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                firstName:
                  type: "string"
                  description: "User's first name"
                lastName:
                  type: "string"
                  description: "User's last name"
                bio:
                  type: "string"
                  description: "User's bio"
      responses:
        "200":
          description: "Returns the current user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: "Unauthorized - Invalid or missing authentication token"
  /user/resend-verification-email:
    get:
      description: "Resend the verification email"
      tags:
        - "User"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Returns email"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  email:
                    type: "string"
  /user/update-password:
    post:
      description: "Update the current user's password"
      tags:
        - "User"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Returns success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  success:
                    type: "boolean"
  /user/enable-mfa:
    post:
      description: "Enable MFA for the current user"
      tags:
        - "User"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Returns the current user"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  uri:
                    type: "string"
  /user/confirm-mfa:
    post:
      description: "Confirm MFA"
      tags:
        - "User"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "token"
              properties:
                token:
                  type: "string"
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: "Returns success status"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  success:
                    type: "boolean"
        "401":
          description: "Unauthorized - Invalid MFA code or MFA not enabled"
  /user/disable-mfa:
    post:
      description: "Disable MFA"
      tags:
        - "User"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Returns success status"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  success:
                    type: "boolean"
        "401":
          description: "Unauthorized - Invalid MFA code or MFA not enabled"
...
components:
  schemas:
...
    User:
      properties:
        id:
          type: "string"
          format: "uuid"
        email:
          type: "string"
          format: "email"
        firstName:
          type: "string"
        lastName:
          type: "string"
        bio:
          type: "string"
        emailVerified:
          type: "boolean"
        lastLoginAt:
          type: "string"
          format: "date-time"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
...